........
  r30232 | thebolt00 | 2008-07-07 20:57:56 +0200 (Mon, 07 Jul 2008) | 2 lines
  
  Killed of a number of warnings
........


git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/branches/release/V1.4@30866 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ include/csgfx/vertexlistwalker.h
@@ -161,7 +161,7 @@ private:
         double orgVal = double (*(C*)data);
         if (Signed)
         {
-          orgVal = (orgVal + (-range - 1)) / double (range*2+1);
+          orgVal = (orgVal + (-range - 1)) / double ((int64)range*2+1);
           newComp = -1.0 + orgVal * 2.0;
         }
         else
